refactor: split CloudSyncIntent out of the coordinator file
The file crossed SwiftLint's 400-line limit in 273df64e and this branch pushed
it to 418. The seam was already there: the intent vocabulary and run outcomes
are values four other files reason about without touching the serialization,
so they move to CloudSyncIntent.swift (110 lines) and the coordinator keeps
the machinery (318). No behavior change.
